wip: passing app ptrs across jni boundary

+++ gs-jni/src/main/java/dev/nateschieber/groovesprings/jni/Vst3AudioHostAppPtr.java
@@ -1,8 +1,8 @@
 package dev.nateschieber.groovesprings.jni;
 
 public class Vst3AudioHostAppPtr {
-    public long address;
+    public int address;
 
-    public Vst3AudioHostAppPtr(long address) {this.address = address;}
-    public long getAddress() { return address; }
+    public Vst3AudioHostAppPtr(int address) {this.address = address;}
+    public int getAddress() { return address; }
 }
